Normalize line endings to LF
