+++ soil_id/tests/us/test_us.py
@@ -20,22 +20,22 @@ from soil_id.us_soil import list_soils, rank_soils
 
 test_locations = [
     {"lon": -121.5111084, "lat": 45.6508331},
-    # {"lon": -101.9733687, "lat": 33.81246789},
-    # {"lon": -121.0347381, "lat": 45.88932423},
-    # {"lon": -85.50621214, "lat": 39.26009312},
-    # {"lon": -94.31005777, "lat": 42.63413723},
-    # {"lon": -99.55016693, "lat": 37.48216451},
-    # {"lon": -157.2767099, "lat": 62.32776717},
-    # {"lon": -156.4422738, "lat": 63.52666854},
-    # {"lon": -119.4596489, "lat": 43.06450312},
-    # {"lon": -69.28246582, "lat": 47.21392200},
-    # {"lon": -158.4018264, "lat": 60.42282639},
-    # {"lon": -121.8166, "lat": 48.6956},
-    # {"lat": 34.92816, "lon": -114.80764},  # NOTCOM
-    # {"lat": 35.599180, "lon": -120.491439},  # previous crash: no objects to concatenate
-    # {"lon": -122.084000, "lat": 37.422000},  # missing LCC
-    # {"lat": 42.494912, "lon": -123.064531},  # crash: could not broadcast input array
-    # {"lat": 40.79861, "lon": -112.35477},  # crash: str object has no attribute rank_data_csv
+    {"lon": -101.9733687, "lat": 33.81246789},
+    {"lon": -121.0347381, "lat": 45.88932423},
+    {"lon": -85.50621214, "lat": 39.26009312},
+    {"lon": -94.31005777, "lat": 42.63413723},
+    {"lon": -99.55016693, "lat": 37.48216451},
+    {"lon": -157.2767099, "lat": 62.32776717},
+    {"lon": -156.4422738, "lat": 63.52666854},
+    {"lon": -119.4596489, "lat": 43.06450312},
+    {"lon": -69.28246582, "lat": 47.21392200},
+    {"lon": -158.4018264, "lat": 60.42282639},
+    {"lon": -121.8166, "lat": 48.6956},
+    {"lat": 34.92816, "lon": -114.80764},  # NOTCOM
+    {"lat": 35.599180, "lon": -120.491439},  # previous crash: no objects to concatenate
+    {"lon": -122.084000, "lat": 37.422000},  # missing LCC
+    {"lat": 42.494912, "lon": -123.064531},  # crash: could not broadcast input array
+    {"lat": 40.79861, "lon": -112.35477},  # crash: str object has no attribute rank_data_csv
 ]
